+++ backend/mk-gateway/src/main/java/de/egladil/web/mk_gateway/infrastructure/persistence/entities/PersistenteTeilnahme.java
@@ -22,12 +22,17 @@ import de.egladil.web.mk_gateway.domain.teilnahmen.Teilnahmeart;
 @Table(name = "TEILNAHMEN")
 @NamedQueries({ @NamedQuery(
 	name = "PersistenteTeilnahme.FIND_BY_NUMMER",
-	query = "select t from PersistenteTeilnahme t where t.teilnahmenummer = :teilnahmenummer order by t.wettbewerbUUID") })
+	query = "select t from PersistenteTeilnahme t where t.teilnahmenummer = :teilnahmenummer order by t.wettbewerbUUID"),
+	@NamedQuery(
+		name = "PersistenteTeilnahme.FIND_BY_WETTBEWERB_ID",
+		query = "select t from PersistenteTeilnahme t where t.wettbewerbUUID = :wettbewerbUUID order by t.uuid") })
 public class PersistenteTeilnahme extends ConcurrencySafeEntity {
 
 	private static final long serialVersionUID = -226547081588693602L;
 
 	public static final String FIND_BY_NUMMER = "PersistenteTeilnahme.FIND_BY_NUMMER";
+
+	public static final String FIND_BY_WETTBEWERB_ID = "PersistenteTeilnahme.FIND_BY_WETTBEWERB_ID";
 
 	@Column(name = "TEILNAHMEART")
 	@Enumerated(EnumType.STRING)