+++ frontend/minikaenguru-ws/apps/mk-admin-app/src/environments/environment.prod.ts
@@ -4,7 +4,7 @@ export const environment = {
 	assetsUrl: 'https://mathe-jung-alt.de/mk-admin-app/assets',
     apiUrl: 'https://mathe-jung-alt.de/mk-gateway/admin',
     storageKeyPrefix: 'mka_',
-	version: '1.3.1',
+	version: '1.3.2',
 	consoleLogActive: false,
 	serverLogActive: true,
 	loglevel: 4